Change Link Component to ExternalLink
The purpose of the Link component was to stop having to copy/paste the `target="_blank" and rel="noopener noreferrer"` attributes on external links, as well as to add an SVG to indicate that the link is external. Normal links were not given any special styling by the Link component. Thus, normal links can now continue to use the regular HTML anchor tag, and external links can use the Link component, and so Link was renamed ExternalLink to better reflect its purpose.