Remove incorrect usages of "isExternal" parameter
Since ExternalLink used to just be Link, it used to accept and "isExternal" parameter. Fixed various usages of this component that still passed an isExternal parameter to the component.

The ExternalLink component uses a spread operator on its parameters to allow for flexibility, so isExternal was being displayed as a property of the resulting HTML tag anywhere isExternal had been passed into it, which caused HTML validation errors. This is now fixed.